+++ sqlitecluster/SQLiteServer.h
@@ -1,12 +1,14 @@
 #pragma once
 class SQLiteCommand;
 
+#include <libstuff/STCPmanager.h>
+
 // This is an abstract class to define the interface for a "server" that a SQLiteNode can communicate with to process
 // commands it receives.
-class SQLiteServer : public STCPServer {
+class SQLiteServer : public STCPManager {
   public:
     // Constructor initializes underlying STCPServer.
-    SQLiteServer(const string& host) : STCPServer(host) { }
+    SQLiteServer() : STCPManager() { }
 
     // An SQLiteNode will call this to pass a command to a server for processing. The isNew flag is set if this is the
     // first time this command has been sent to this server, as opposed to being an existing command, such as one that
