+++ test/tests/SQLiteNodeTest.cpp
@@ -20,7 +20,7 @@ class SQLiteNodeTester {
 
 class TestServer : public SQLiteServer {
   public:
-    TestServer(const string& host) : SQLiteServer(host) { }
+    TestServer() : SQLiteServer() { }
 
     virtual void acceptCommand(unique_ptr<SQLiteCommand>&& command, bool isNew) { }
     virtual void cancelCommand(const string& commandID) { }
@@ -48,7 +48,7 @@ struct SQLiteNodeTest : tpunit::TestFixture {
         int fd = mkstemp(filename);
         close(fd);
         shared_ptr<SQLitePool> dbPool = make_shared<SQLitePool>(10, filename, 1000000, 5000, 0);
-        TestServer server("");
+        TestServer server;
         string peerList = "host1.fake:15555?nodeName=peer1,host2.fake:16666?nodeName=peer2,host3.fake:17777?nodeName=peer3,host4.fake:18888?nodeName=peer4";
         SQLiteNode testNode(server, dbPool, "test", "localhost:19998", peerList, 1, 1000000000, "1.0");
